fix: error 로직 분리 -> 순환참조 해결
